vertex-ai-deployment/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
diffusers>=0.29.0
transformers>=4.30.0
accelerate>=0.20.0
xformers>=0.0.20
pillow>=9.0.0
flask>=2.3.0
google-cloud-storage>=2.10.0
huggingface_hub>=0.19.0